moved most of App.apply to Wiring class (as in easy-download)
